requirements.txt
streamlit
sqlite-utils
python-dotenv